refactor: 도서별 토론방 조회 API를 BookController로 이동
