+++ backend/src/main/java/todoktodok/backend/book/presentation/BookController.java
@@ -6,6 +6,7 @@ import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -16,6 +17,8 @@ import todoktodok.backend.book.application.dto.response.AladinBookResponse;
 import todoktodok.backend.book.application.dto.response.LatestAladinBookPageResponse;
 import todoktodok.backend.book.application.service.command.BookCommandService;
 import todoktodok.backend.book.application.service.query.BookQueryService;
+import todoktodok.backend.discussion.application.dto.response.LatestDiscussionPageResponse;
+import todoktodok.backend.discussion.application.service.query.DiscussionQueryService;
 import todoktodok.backend.global.auth.Auth;
 import todoktodok.backend.global.auth.Role;
 import todoktodok.backend.global.resolver.LoginMember;
@@ -27,6 +30,7 @@ public class BookController implements BookApiDocs {
 
     private final BookCommandService bookCommandService;
     private final BookQueryService bookQueryService;
+    private final DiscussionQueryService discussionQueryService;
 
     @Auth(value = Role.USER)
     @PostMapping
@@ -58,5 +62,17 @@ public class BookController implements BookApiDocs {
     ) {
         return ResponseEntity.status(HttpStatus.OK)
                 .body(bookQueryService.searchByPaging(size, cursor, keyword));
+    }
+
+    @Auth(value = Role.USER)
+    @GetMapping("/{bookId}/discussions")
+    public ResponseEntity<LatestDiscussionPageResponse> getDiscussionsByBook(
+            @LoginMember final Long memberId,
+            @PathVariable final Long bookId,
+            @RequestParam final int size,
+            @RequestParam(required = false) final String cursor
+    ) {
+        return ResponseEntity.status(HttpStatus.OK)
+                .body(discussionQueryService.getDiscussionsByBook(memberId, bookId, size, cursor));
     }
 }
